+++ omop_alchemy/cdm/model/vocabulary/concept.py
@@ -34,6 +34,13 @@ class Concept(
         omop_index(__tablename__, "vocabulary_id"),
         omop_index(__tablename__, "domain_id"),
         omop_index(__tablename__, "concept_class_id"),
+        # Has to be wrapped in func.lower() as that is the common query
+        # as it prevents captialisation mismatches between query and data.
+        omop_index(
+            __tablename__,
+            sa.func.lower(sa.column("concept_name")),
+            name="ix_concept_concept_name_lower",
+        ),
         omop_table_options(cluster_on=omop_primary_key_index_name("concept")),
     )
     concept_id: so.Mapped[int] = so.mapped_column(primary_key=True)